remove unused utils, convert utils to TS
