--- client/src/Utils/ReadMe.md
@@ -1,1 +1,0 @@
-#Utils folder
