+++ client/src/main.tsx
@@ -5,7 +5,7 @@ import { BrowserRouter as Router } from "react-router-dom";
 import { Provider } from "react-redux";
 import { persistor, store } from "@/store.js";
 import { PersistGate } from "redux-persist/integration/react";
-import I18nLoader from "./Components/v1/I18nLoader/index.jsx";
+import I18nLoader from "./Components/v2/i18nLoader";
 import { initApiClient } from "./Utils/ApiClient.js";
 
 initApiClient(store);
